requirements.txt
python-telegram-bot
anthropic
openai